client: Actually use WindowIdentifier from ashpd
As we already re-implement the SecretProxy + SecretInterface in the
portal crate
In the future, we would be controlling both the client and the server
side allowing us
to easily to adapt to changes in the secret portal
