+++ Cargo.toml
@@ -20,6 +20,7 @@ rust-version = "1.80"
 exclude = ["org.freedesktop.Secrets.xml"]
 
 [workspace.dependencies]
+ashpd = {version = "0.9", git = "https://github.com/bilelmoussaoui/ashpd", default-features = false}
 endi = "1.1"
 clap = { version = "4.4.6", features = [ "cargo", "derive" ] }
 futures-channel = "0.3"
