+++ client/Cargo.toml
@@ -14,10 +14,10 @@ version.workspace = true
 
 [dependencies]
 aes = { version = "0.8", features = ["zeroize"], optional = true }
+ashpd.workspace = true
 async-fs = { version = "2.1.0", optional = true }
 async-io = { version = "2.3.4", optional = true }
 async-lock = { version = "3.2.0", optional = true }
-async-net = { version = "2.0.0", optional = true }
 blocking = { version = "1.5.1", optional = true }
 cbc = { version = "0.1", features = ["zeroize"], optional = true }
 cipher = { version = "0.4", features = [
@@ -65,11 +65,11 @@ async-std = [
     "dep:async-fs",
     "dep:async-io",
     "dep:async-lock",
-    "dep:async-net",
     "dep:blocking",
     "dep:futures-lite",
+    "ashpd/async-std",
 ]
-tokio = ["zbus/tokio", "dep:tokio"]
+tokio = ["zbus/tokio", "dep:tokio", "ashpd/tokio"]
 native_crypto = [
     "dep:aes",
     "dep:cbc",
@@ -83,6 +83,7 @@ native_crypto = [
     "dep:subtle",
 ]
 openssl_crypto = ["dep:openssl"]
+tracing = ["dep:tracing", "ashpd/tracing"]
 
 [package.metadata.docs.rs]
 features = ["unstable"]
