+++ client/src/dbus/api/mod.rs
@@ -21,7 +21,6 @@ mod session;
 pub use collection::Collection;
 pub use item::Item;
 pub(crate) use prompt::Prompt;
-pub use prompt::WindowIdentifier;
 #[cfg(not(feature = "unstable"))]
 pub(crate) use properties::Properties;
 #[cfg(feature = "unstable")]
