+++ client/src/dbus/error.rs
@@ -31,8 +31,6 @@ pub enum Error {
     NotFound(String),
     /// Input/Output.
     IO(std::io::Error),
-    /// Invalid window identifier
-    InvalidWindowIdentifier(String),
 }
 
 impl From<zbus::Error> for Error {
@@ -76,9 +74,6 @@ impl fmt::Display for Error {
             Self::Deleted => write!(f, "Item/Collection was deleted, can no longer be used"),
             Self::NotFound(name) => write!(f, "The collection '{name}' doesn't exists"),
             Self::Dismissed => write!(f, "Prompt was dismissed"),
-            Self::InvalidWindowIdentifier(identifier) => {
-                write!(f, "Invalid window identifier {identifier}")
-            }
         }
     }
 }
