+++ client/src/dbus/service.rs
@@ -1,12 +1,10 @@
 use std::sync::Arc;
 
+use ashpd::WindowIdentifier;
 use futures_util::{Stream, StreamExt};
 use zbus::zvariant::OwnedObjectPath;
 
-use super::{
-    api::{self, WindowIdentifier},
-    Algorithm, Collection, Error, DEFAULT_COLLECTION,
-};
+use super::{api, Algorithm, Collection, Error, DEFAULT_COLLECTION};
 use crate::Key;
 
 /// The entry point of communicating with a [`org.freedesktop.Secrets`](https://specifications.freedesktop.org/secret-service-spec/latest/index.html) implementation.
